requirements.txt
streamlit
torch
numpy
joblib
scikit-learn
pandas
fpdf
plotly
Pillow